requirements.txt
Flask==2.3.2
APScheduler==3.10.1
feedparser==6.0.11
python-dotenv==1.0.1